7/25

+++ demo/src/main/java/com/example/demo/Controller/CategoryController.java
@@ -5,7 +5,6 @@ import com.example.demo.Service.CategoryService;
 import com.example.demo.Service.PostCommentsViewService;
 import com.example.demo.Service.SubCategoryService;
 import com.example.demo.Util.ApiResponse;
-import com.example.demo.Util.RedisCache;
 import com.example.demo.Validator.ValidString;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -38,22 +37,22 @@ public class CategoryController {
 
     //------------------------------------------------------------------------------------------
     @GetMapping("/category/sub_category")
-    public ResponseEntity<ApiResponse<Map<Integer, Map<String, List<SubCategory>>>>> getAllSubCategories() throws JsonProcessingException {
+    public ResponseEntity<ApiResponse<Map<Byte, Map<String, List<SubCategory>>>>> getAllSubCategories() throws JsonProcessingException {
         Jedis jedis = new Jedis("localhost");
         boolean existsInCache = jedis.exists(CATEGORY_SUB_CATEGORY_CACHE_KEY);
         if(existsInCache){
             logger.info("CATEGORY_SUB_CATEGORY_CACHE exists in Redis cache:::");
             logger.info("Get CATEGORY_SUB_CATEGORY from Redis cache:::");
             String json = jedis.get(CATEGORY_SUB_CATEGORY_CACHE_KEY);
-            Map<Integer, Map<String, List<SubCategory>>> res = objectMapper.readValue(json, new TypeReference<Map<Integer, Map<String, List<SubCategory>>>>(){});
-            ApiResponse<Map<Integer, Map<String, List<SubCategory>>>> apiResponse = ApiResponse.success(res);
+            Map<Byte, Map<String, List<SubCategory>>> res = objectMapper.readValue(json, new TypeReference<Map<Byte, Map<String, List<SubCategory>>>>(){});
+            ApiResponse<Map<Byte, Map<String, List<SubCategory>>>> apiResponse = ApiResponse.success(res);
             return ResponseEntity.ok(apiResponse);
         }else{
             logger.info("CATEGORY_SUB_CATEGORY_CACHE doesn't exist in Redis cache:::");
             logger.info("Get CATEGORY_SUB_CATEGORY from MySQL:::");
-            Map<Integer, Map<String, List<SubCategory>>> res = categoryService.getAllSubCategories();
+            Map<Byte, Map<String, List<SubCategory>>> res = categoryService.getAllSubCategories();
             if(res!=null){
-                ApiResponse<Map<Integer, Map<String, List<SubCategory>>>> apiResponse = ApiResponse.success(res);
+                ApiResponse<Map<Byte, Map<String, List<SubCategory>>>> apiResponse = ApiResponse.success(res);
                 return ResponseEntity.ok(apiResponse);
             }else{
                 ApiResponse errorResponse = ApiResponse.error(500, "Internal Server Error", "Internal Server Error");
@@ -64,7 +63,7 @@ public class CategoryController {
     //------------------------------------------------------------------------------------------
     //show all posts under specific category
     @GetMapping("/categories/{sub_categoryId}")
-    public ResponseEntity<ApiResponse<List<Map<String,Object>>>> PostsCommentsViewWithSubCategoryId(@PathVariable Integer sub_categoryId) throws IOException {
+    public ResponseEntity<ApiResponse<List<Map<String,Object>>>> PostsCommentsViewWithSubCategoryId(@PathVariable Short sub_categoryId) throws IOException {
         if(!ValidString.SubCategoryIdEmpty(sub_categoryId) || !ValidString.SubCategoryIdLength(sub_categoryId)){
             ApiResponse errorResponse =ApiResponse.error(404, "No Such Sub Category", "Not Found");
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponse);
